# backend/golang/internal/api/docs/swagger.yaml
definitions:
...
  dtos.Customer:
    properties:
      birthdate:
        type: string
      cpf:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  dtos.CustomerInfoBody:
    properties:
      birthdate:
        type: string
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  dtos.Service:
    properties:
      duration:
        type: integer
      id:
        type: string
      price:
        type: integer
      serviceName:
        type: string
      serviceNameId:
        type: string
      specialistId:
        type: string
      specialistName:
        type: string
      specialization:
        type: string
      specializationId:
        type: string
    type: object
  dtos.ServiceInfoBody:
    properties:
      duration:
        description: in minutes
        type: integer
      price:
        type: integer
      serviceNameId:
        type: string
      specialistId:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /customers:
    get:
      description: Get a page of customers with filters
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Name
        in: query
        name: name
        type: string
      - description: CPF
        in: query
        name: cpf
        type: string
      - description: Phone
        in: query
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.Customer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List customers
      tags:
      - Customers
    post:
      description: Create an new customer
      parameters:
      - description: Customer information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dtos.CustomerInfoBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Customer'
      security:
      - ApiKeyAuth: []
      summary: Create customer
      tags:
      - Customers
  /customers/{customer_id}:
    delete:
      description: Delete a customer by id
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete customer
      tags:
      - Customers
    get:
      description: Get a customer by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Customer'
      security:
      - ApiKeyAuth: []
      summary: Get customer
      tags:
      - Customers
    put:
      description: Update a customer by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Customer'
      security:
      - ApiKeyAuth: []
      summary: Update customer
      tags:
      - Customers
  /customers/count:
    get:
      description: Count customers with filters
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: CPF
        in: query
        name: cpf
        type: string
      - description: Phone
        in: query
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Count customers
      tags:
      - Customers
...
  /services:
    get:
      description: Get a page of services with filters
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Service
        in: query
        name: service
        type: string
      - description: Specialist
        in: query
        name: specialist
        type: string
      - description: Specialization
        in: query
        name: specialization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.Service'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List services
      tags:
      - Services
    post:
      description: Create an new service
      parameters:
      - description: Service information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dtos.ServiceInfoBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Service'
      security:
      - ApiKeyAuth: []
      summary: Create service
      tags:
      - Services
  /services/{service_id}:
    delete:
      description: Delete an service by id
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete service
      tags:
      - Services
    get:
      description: Get a service by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Service'
      security:
      - ApiKeyAuth: []
      summary: Get service
      tags:
      - Services
    put:
      description: Create an new service
      parameters:
      - description: Service information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dtos.ServiceInfoBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Service'
      security:
      - ApiKeyAuth: []
      summary: Create service
      tags:
      - Services
  /services/count:
    get:
      description: Count services with filters
      parameters:
      - description: Service
        in: query
        name: service
        type: string
      - description: Specialist
        in: query
        name: specialist
        type: string
      - description: Specialization
        in: query
        name: specialization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      security:
      - ApiKeyAuth: []
      summary: Count services
      tags:
      - Services
...
swagger: "2.0"